lab-03

+++ src/lab_02/SuggestWeight.java
@@ -3,6 +3,7 @@ package lab_02;
 import java.util.Scanner;
 
 public class SuggestWeight {
+
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
         System.out.print("Please input your height: ");
@@ -16,7 +17,7 @@ public class SuggestWeight {
 
         if (BMI < 18.5) {
             System.out.println("Underweight");
-            adjustWeight = (float) (18.5 - BMI) * (height * 2);
+            adjustWeight =  (18.5f - BMI) * (height * 2);
             float newWeight = weight + adjustWeight;
             System.out.println("You should increase: " + adjustWeight + "kg");
             System.out.println("Your weight should be: " + newWeight);
@@ -28,7 +29,7 @@ public class SuggestWeight {
             } else {
                 System.out.println("Obesity");
             }
-            adjustWeight = (float) (BMI - 24.9) * (height * 2);
+            adjustWeight = (BMI - 24.9f) * (height * 2);
             float newWeight = weight - adjustWeight;
             System.out.println("You should decrease: " + adjustWeight + "kg");
             System.out.println("Your weight should be: " + newWeight);
